Confirmed working version
With a major thanks to: GohanCZ who provided a save with a weapon with armor parts I was able to fully test and refine the script for least impact. The part replacement is instantaneous it is likely the player will never see it. I commented out all debug lines except Destroy Triggered, and Parts Destroyed that way in the future if someone reports changing weapon parts we know if this script is the cause.