BaS 9-29-22 Hotfix
Compatibility for GAMMA and BaS 9-29-22. No crashes, guns are NOT added to loot at this time. Most guns are added to partlist, all model addresses are updated.

Remove mod 77 and 78.
Turn off Part Type Fixer as some of the upgraded guns are missing parts still, but ALL base guns, and most kits have parts.

Balance and final work to be completed over time.