To prevent stutters, always resupply traders no matter the distance
There is a stuttering issue caused by many items being spawned on a trader when the player gets close enough. The change allows these items to be spawned in as soon as possible rather than wait for the player to get close.

The stuttering is not so much the act of spawning the items, but the need to load in game assets at the time of spawn. If the assets have been loaded, then there wouldn't be a stutter when the items get spawned.